Reorgnaise damlc to put the test stuff in a separate module
